// cosmowcl/src/app/shared/models/api/cosmoauthorization/gruppo.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cw-profilazione
 * API dei servizi di profilazione usati dal common workspace
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RiferimentoEnte } from './riferimentoEnte.d';
import { RiferimentoUtente } from './riferimentoUtente.d';
import { TipoPratica } from './tipoPratica.d';


/**
 * Gruppo che puo\' essere associato a un utente
 */
export interface Gruppo { 
    id: number;
    codice: string;
    nome: string;
    descrizione?: string;
    ente: RiferimentoEnte;
    utenti?: Array<RiferimentoUtente>;
    tipologiePratiche?: Array<TipoPratica>;
}